Avoid using RTTI for list tests when possible
With Clang, GCC and MSVC, we can generate the typename using compiler
macros and some string manipulation. Where this is possible, we will use
this instead of RTTI. Where it is not, we will fall back to RTTI and/or
index-based default typename string generation.

Special thanks to Corentin Schreiber (cschreib@strw.leidenuniv.nl) for
the inspiration here.